mosquitto.sh run native mosquitto in foreground
The goal of mosquitto.sh is to get a mosquitto service up and running so
that tests can built and run in the native OS environment.

A mosquitto container running on a MacOS host cannot share a unix socket
from the Linux container to the MacOS host.

Running a native mosquitto instance from mosquitto.sh solves this
problem. The native mosquitto process can create/listen a
host-compatible unix socket.

The script is modified to run a locally installed mosquitto by default,
if available, and fallback to running mosquitto in a container.

For both the locally installed and containerized paths, the script now
just runs in the foreground; no more start/stop/status subcommands.